-- Fixed minor bugs in sacct and sstat commands

+++ branches/slurm-2.1/src/sstat/print.c
@@ -1,5 +1,5 @@
 /*****************************************************************************\
- *  print.c - print functions for sacct
+ *  print.c - print functions for sstat
  *
  *  $Id: print.c 7541 2006-03-18 01:44:58Z da $
  *****************************************************************************
@@ -93,7 +93,7 @@ void print_fields(jobacct_step_rec_t *step)
 		switch(field->type) {
 		case PRINT_AVECPU:
 
-			tmp_char = _elapsed_time((int)step->sacct.ave_cpu, 0);
+			tmp_char = _elapsed_time((long)step->sacct.ave_cpu, 0);
 
 			field->print_routine(field,
 					     tmp_char,
@@ -206,7 +206,7 @@ void print_fields(jobacct_step_rec_t *step)
 					     (curr_inx == field_count));
 			break;
 		case PRINT_MINCPU:
-			tmp_char = _elapsed_time((int)step->sacct.min_cpu, 0);
+			tmp_char = _elapsed_time((long)step->sacct.min_cpu, 0);
 			field->print_routine(field,
 					     tmp_char,
 					     (curr_inx == field_count));
